% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse_load.R
\name{gse_load}
\alias{gse_load}
\title{gse_load}
\usage{
gse_load(dir_path, files = NULL, sample_type = NULL)
}
\arguments{
\item{dir_path}{char, path to folder containing the gsm files (in a *.tar.gz format)}

\item{files}{c(chars), name of each gsm (for example, gsm00223)to extract into the seurat object,
otherwise it'll take all of the gsm files in the folder and compile them.}

\item{sample_type}{c(chars), the grouping of each sample.  For example,}
}
\value{
The seurat object
}
\description{
Takes a folder_path containing a set of pattern "*.tar.gz",
and returns one them into one seurat object.
}
