% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlplate.R
\name{read_xlplate}
\alias{read_xlplate}
\title{read_xlplate}
\usage{
read_xlplate(excel_file_path, sheet_name)
}
\arguments{
\item{excel_file_path}{The file path to the excel file}

\item{sheet_name}{The sheet name to use the read_plate() function on.}
}
\value{
Returns a tibble containing the plate variables.
}
\description{
Allows the plater::read_plate() function to look into excel sheets.
}
